Insert base type for the dimension system type for GNAT GPL 2017
Improvements in the dimension system in GNAT GPL 2017 mean that some
previously valid type conversions are now seen as invalid because the
dimensions are not forgotten anymore. Insert a type Base_Unit_Type for
the basis of Unit_Type, so that conversions to Base_Unit_Type remove
the dimensions, while conversions to Unit_Type (the root of the dimension
system) preserve dimensions.

The resulting code is compilable both with GNAT GPL 2016, and will be
compilable with GNAT GPL 2017 with an adapted runtime. (The current
runtime is specific to GNAT GPL 2016.)